Rename favorites button file
