+++ resources/js/components/frontend/nav/favorites.tsx
